refactor: extract board watchers module
