+++ packages/board/server.mjs
@@ -88,86 +88,9 @@ import {
 } from './lib/fleet.mjs';
 import { readAlertsFired, writeAlertsFired, fireEmailAlert, firePushAlert, startAlertCron } from './lib/alerts.mjs';
 import { getResume, shareStatePath, getShareState, saveShareState, publishReport, toggleShare, generateShareHTML } from './lib/share.mjs';
+import { watchBeads, watchVerdicts } from './lib/watchers.mjs';
 
 // ── Helpers ────────────────────────────────────────────────────────────────────
-
-// ── File watcher ───────────────────────────────────────────────────────────────
-function watchBeads() {
-  // Watch every registered project's beads files.
-  // Note: bd create only writes to dolt DB, NOT interactions.jsonl. So we must
-  // watch BOTH: (a) interactions.jsonl for status/priority changes (from bd
-  // update/close), and (b) the dolt manifest/journal for new-issue detection.
-  const projects = listProjects();
-  const dirs = projects.map(p => p.path);
-  if (!dirs.includes(process.cwd())) dirs.push(process.cwd());
-
-  const broadcast = (dir) => {
-    bdCacheInvalidate(dir);
-    for (const res of sseClients) {
-      if (res._gctoCwd === dir) {
-        try {
-          res.write(`event: tasks\ndata: ${JSON.stringify(getTasks(dir))}\n\n`);
-          res.write(`event: pipeline\ndata: ${JSON.stringify(getPipeline(dir))}\n\n`);
-          res.write(`event: inbox\ndata: ${JSON.stringify(getInbox(dir))}\n\n`);
-        } catch {}
-      }
-    }
-  };
-
-  // Debounce per-dir: dolt writes can fire 3-5 events in <50ms during a single
-  // bd command. Collapse them into one broadcast 200ms after the last event.
-  const debouncers = new Map();
-  const schedule = (dir) => {
-    if (debouncers.has(dir)) clearTimeout(debouncers.get(dir));
-    debouncers.set(dir, setTimeout(() => {
-      debouncers.delete(dir);
-      broadcast(dir);
-    }, 200));
-  };
-
-  for (const dir of dirs) {
-    // (a) interactions.jsonl — captures bd update/close
-    const interactionsFile = path.join(dir, '.beads', 'interactions.jsonl');
-    if (fs.existsSync(interactionsFile)) {
-      try { fs.watch(interactionsFile, () => schedule(dir)); } catch {}
-    }
-    // (b) dolt embeddeddolt directory (recursive) — captures bd create
-    const doltDir = path.join(dir, '.beads', 'embeddeddolt');
-    if (fs.existsSync(doltDir)) {
-      try { fs.watch(doltDir, { recursive: true }, () => schedule(dir)); } catch {}
-    }
-  }
-}
-
-// Watch ~/.great_cto/verdicts/ — push pipeline updates whenever an agent
-// emits a verdict (any project gets the broadcast for its own cwd).
-function watchVerdicts() {
-  const verdictDir = path.join(GREAT_CTO_DIR, 'verdicts');
-  if (!fs.existsSync(verdictDir)) {
-    try { fs.mkdirSync(verdictDir, { recursive: true }); } catch { return; }
-  }
-  let pushTimer = null;
-  const broadcastPipeline = () => {
-    if (pushTimer) clearTimeout(pushTimer);
-    // debounce: collapse a burst of writes (multiple agents finishing within ~150ms)
-    pushTimer = setTimeout(() => {
-      for (const res of sseClients) {
-        const dir = res._gctoCwd || process.cwd();
-        try {
-          res.write(`event: pipeline\ndata: ${JSON.stringify(getPipeline(dir))}\n\n`);
-          res.write(`event: inbox\ndata: ${JSON.stringify(getInbox(dir))}\n\n`);
-        } catch { sseClients.delete(res); }
-      }
-    }, 150);
-  };
-  try {
-    fs.watch(verdictDir, () => broadcastPipeline());
-    // Also watch each existing log file (some agents append to existing)
-    for (const f of fs.readdirSync(verdictDir).filter(x => x.endsWith('.log'))) {
-      try { fs.watch(path.join(verdictDir, f), () => broadcastPipeline()); } catch {}
-    }
-  } catch {}
-}
 
 // ── HTTP router ────────────────────────────────────────────────────────────────
 const MAX_BODY_BYTES = 2 * 1024 * 1024; // 2 MB — board payloads are tiny; cap guards against unbounded `body += c` accumulation.
